requirements.txt
requests
aiofiles
ujson
python-multipart
graphql-core < 3
graphene
isort

# Testing & Linting
pytest
pytest-cov
pytest-timeout
black